fix float64 for tasks 3 4 5

+++ math/convolutions_and_pooling/3-convolve_grayscale.py
@@ -15,13 +15,14 @@ def convolve_grayscale(images, kernel, padding='same', stride=(1, 1)):
         ph, pw = 0, 0
     else:
         ph, pw = padding
-    padded = np.pad(images, ((0, 0), (ph, ph), (pw, pw)), mode='constant')
+    padded = np.pad(images, ((0, 0), (ph, ph), (pw, pw)),
+                    mode='constant').astype('float64')
     oh = (h + 2 * ph - kh) // sh + 1
     ow = (w + 2 * pw - kw) // sw + 1
     output = np.zeros((m, oh, ow))
     for i in range(oh):
         for j in range(ow):
             output[:, i, j] = np.sum(
-                padded[:, i * sh:i * sh + kh, j * sw:j * sw + kw] * kernel,
-                axis=(1, 2))
+                padded[:, i * sh:i * sh + kh, j * sw:j * sw + kw] *
+                kernel.astype('float64'), axis=(1, 2))
     return output
